feat: adds worker messages
