+++ horde/apis/v2/__init__.py
@@ -4,7 +4,10 @@
 
 import horde.apis.v2.base as base
 import horde.apis.v2.kobold as kobold
+import horde.apis.v2.kobold_styles as kobold_styles
 import horde.apis.v2.stable as stable
+import horde.apis.v2.stable_styles as stable_styles
+import horde.apis.v2.styles as styles
 from horde.apis.v2.base import api
 
 api.add_resource(stable.ImageAsyncGenerate, "/generate/async")
@@ -13,21 +16,21 @@ api.add_resource(stable.ImageAsyncCheck, "/generate/check/<string:id>")
 api.add_resource(stable.Aesthetics, "/generate/rate/<string:id>")
 api.add_resource(stable.ImageJobPop, "/generate/pop")
 api.add_resource(stable.ImageJobSubmit, "/generate/submit")
-api.add_resource(stable.ImageStyle, "/styles/image")
-api.add_resource(stable.SingleImageStyle, "/styles/image/<string:style_id>")
-api.add_resource(stable.SingleImageStyleByName, "/styles/image_by_name/<string:style_name>")
-api.add_resource(stable.ImageStyleExample, "/styles/image/<string:style_id>/example")
-api.add_resource(stable.SingleImageStyleExample, "/styles/image/<string:style_id>/example/<string:example_id>")
+api.add_resource(stable_styles.ImageStyle, "/styles/image")
+api.add_resource(stable_styles.SingleImageStyle, "/styles/image/<string:style_id>")
+api.add_resource(stable_styles.SingleImageStyleByName, "/styles/image_by_name/<string:style_name>")
+api.add_resource(stable_styles.ImageStyleExample, "/styles/image/<string:style_id>/example")
+api.add_resource(stable_styles.SingleImageStyleExample, "/styles/image/<string:style_id>/example/<string:example_id>")
 api.add_resource(kobold.TextAsyncGenerate, "/generate/text/async")
 api.add_resource(kobold.TextAsyncStatus, "/generate/text/status/<string:id>")
 api.add_resource(kobold.TextJobPop, "/generate/text/pop")
 api.add_resource(kobold.TextJobSubmit, "/generate/text/submit")
-api.add_resource(kobold.TextStyle, "/styles/text")
-api.add_resource(kobold.SingleTextStyle, "/styles/text/<string:style_id>")
-api.add_resource(kobold.SingleImageStyleByName, "/styles/text_by_name/<string:style_name>")
-api.add_resource(base.Collection, "/collections")
-api.add_resource(base.SingleCollection, "/collections/<string:collection_id>")
-api.add_resource(base.SingleCollectionByName, "/collection_by_name/<string:collection_name>")
+api.add_resource(kobold_styles.TextStyle, "/styles/text")
+api.add_resource(kobold_styles.SingleTextStyle, "/styles/text/<string:style_id>")
+api.add_resource(kobold_styles.SingleImageStyleByName, "/styles/text_by_name/<string:style_name>")
+api.add_resource(styles.Collection, "/collections")
+api.add_resource(styles.SingleCollection, "/collections/<string:collection_id>")
+api.add_resource(styles.SingleCollectionByName, "/collection_by_name/<string:collection_name>")
 api.add_resource(base.Users, "/users")
 api.add_resource(base.UserSingle, "/users/<string:user_id>")
 api.add_resource(base.FindUser, "/find_user")
@@ -35,6 +38,8 @@ api.add_resource(base.SharedKey, "/sharedkeys")
 api.add_resource(base.SharedKeySingle, "/sharedkeys/<string:sharedkey_id>")
 api.add_resource(base.Workers, "/workers")
 api.add_resource(base.WorkerSingle, "/workers/<string:worker_id>")
+api.add_resource(base.WorkerMessages, "/workers/messages")
+api.add_resource(base.SingleWorkerMessage, "/workers/messages/<string:message_id>")
 api.add_resource(base.WorkerSingleName, "/workers/name/<string:worker_name>")
 api.add_resource(base.TransferKudos, "/kudos/transfer")
 api.add_resource(base.AwardKudos, "/kudos/award")
