+++ horde/apis/v2/base.py
@@ -23,11 +23,10 @@ from horde.argparser import args
 from horde.classes.base import settings
 from horde.classes.base.detection import Filter
 from horde.classes.base.news import News
-from horde.classes.base.style import StyleCollection
 from horde.classes.base.team import Team, find_team_by_id, get_all_teams
 from horde.classes.base.user import User, UserSharedKey
 from horde.classes.base.waiting_prompt import WaitingPrompt
-from horde.classes.base.worker import Worker
+from horde.classes.base.worker import Worker, WorkerMessage
 from horde.consts import HORDE_VERSION
 from horde.countermeasures import CounterMeasures
 from horde.database import functions as database
@@ -41,7 +40,7 @@ from horde.metrics import waitress_metrics
 from horde.patreon import patrons
 from horde.r2 import upload_prompt
 from horde.suspicions import Suspicions
-from horde.utils import ensure_clean, hash_api_key, hash_dictionary, is_profane, sanitize_string
+from horde.utils import hash_api_key, hash_dictionary, is_profane, sanitize_string
 from horde.vars import horde_contact_email, horde_title, horde_url
 
 # Not used yet
@@ -491,6 +490,7 @@ class JobPopTemplate(Resource):
                 if not wp.needs_gen():  # this says if < 1
                     continue
                 worker_ret = self.start_worker(wp)
+                worker_ret["messages"] = database.get_all_active_worker_messages(self.worker.id)
                 # logger.debug(worker_ret)
                 if worker_ret is None:
                     continue
@@ -504,7 +504,7 @@ class JobPopTemplate(Resource):
         if self.worker.maintenance:
             raise e.WorkerMaintenance(self.worker.maintenance_msg)
         # logger.debug(self.skipped)
-        return {"id": None, "ids": [], "skipped": self.skipped}, 200
+        return {"id": None, "ids": [], "skipped": self.skipped, "messages": database.get_all_active_worker_messages(self.worker.id)}, 200
 
     def get_sorted_wp(self, priority_user_ids=None):
         """Extendable class to retrieve the sorted WP list for this worker"""
@@ -875,6 +875,9 @@ class WorkerSingleBase(Resource):
         cached_worker = hr.horde_r_get(cache_name)
         if cache_exists and cached_worker:
             worker_details = json.loads(cached_worker)
+            for msg in worker_details.get("messages", []):
+                msg["created"] = datetime.strptime(msg["created"], "%a, %d %b %Y %H:%M:%S %z")
+                msg["expiry"] = datetime.strptime(msg["expiry"], "%a, %d %b %Y %H:%M:%S %z")
         else:
             worker = database.find_worker_by_id(worker_id)
             if not worker:
@@ -3157,179 +3160,9 @@ class DocsSponsors(Resource):
         return {"html": html_template}, 200
 
 
-## Styles
-
-
-class StyleTemplate(Resource):
-    gentype = "template"
-    args = None
-
-    def get(self):
-        if self.args.sort not in ["popular", "age"]:
-            raise e.BadRequest("'model_state' needs to be one of ['popular', 'age']")
-        styles_ret = database.retrieve_available_styles(
-            style_type=self.gentype,
-            sort=self.args.sort,
-            page=self.args.page - 1,
-            tag=self.args.tag,
-            model=self.args.model,
-        )
-        styles_ret = [st.get_details() for st in styles_ret]
-        return styles_ret, 200
-
-    def post(self):
-        # I have to extract and store them this way, because if I use the defaults
-        # It causes them to be a shared object from the parsers class
-        self.params = {}
-        self.warnings = set()
-        if self.args.params:
-            self.params = self.args.params
-        # For styles, we just store the models in the params
-        self.models = []
-        if self.args.models:
-            self.params["models"] = self.args.models.copy()
-        self.user = None
-        self.validate()
-        return
-
-    def validate(self):
-        self.sharedkey = None
-        if self.args.sharedkey:
-            self.sharedkey = database.find_sharedkey(self.args.sharedkey)
-            if self.sharedkey is None:
-                raise e.BadRequest("This shared key does not exist", "SharedKeyInvalid")
-            shared_key_validity = self.sharedkey.is_valid()
-            if shared_key_validity[0] is False:
-                raise e.BadRequest(shared_key_validity[1], shared_key_validity[2])
-
-
-class SingleStyleTemplateGet(Resource):
-    gentype = "template"
-
-    def get_existing_style(self):
-        if self.existing_style.style_type != self.gentype:
-            raise e.BadRequest(
-                f"Style was found but was of the wrong type: {self.existing_style.style_type} != {self.gentype}",
-                "StyleGetMistmatch",
-            )
-        return self.existing_style.get_details()
-
-    def get_through_id(self, style_id):
-        self.existing_style = database.get_style_by_uuid(style_id, is_collection=False)
-        if not self.existing_style:
-            raise e.ThingNotFound(f"{self.gentype} Style", style_id)
-        return self.get_existing_style()
-
-
-class SingleStyleTemplate(SingleStyleTemplateGet):
-
-    def patch(self, style_id):
-        self.params = {}
-        self.warnings = set()
-        self.args = parsers.style_parser.parse_args()
-        if self.args.params:
-            self.params = self.args.params
-        # For styles, we just store the models in the params
-        self.models = []
-        style_modified = False
-        self.tags = []
-        if self.args.tags:
-            self.tags = self.args.tags.copy()
-            if len(self.tags) > 10:
-                raise e.BadRequest("A style can be tagged a maximum of 10 times.")
-        self.user = database.find_user_by_api_key(self.args["apikey"])
-        if not self.user:
-            raise e.InvalidAPIKey("Style PATCH")
-        self.existing_style = database.get_style_by_uuid(style_id, is_collection=False)
-        if not self.existing_style:
-            raise e.ThingNotFound("Style", style_id)
-        if self.existing_style.user_id != self.user.id:
-            raise e.Forbidden(f"This Style is not owned by user {self.user.get_unique_alias()}")
-        if self.args.models:
-            self.models = self.args.models.copy()
-            if len(self.models) > 5:
-                raise e.BadRequest("A style can only use a maximum of 5 models.")
-            if len(self.models) < 1:
-                raise e.BadRequest("A style has to specify at least one model.")
-        else:
-            self.models = self.existing_style.get_model_names()
-        self.style_name = None
-        if self.args.name:
-            self.style_name = ensure_clean(self.args.name, "style name")
-            style_modified = True
-        self.validate()
-        self.existing_style.name = self.style_name
-        if self.args.info is not None:
-            self.existing_style.info = ensure_clean(self.args.info, "style info")
-            style_modified = True
-        if self.args.public is not None:
-            self.existing_style.public = self.args.public
-            style_modified = True
-        if self.args.nsfw is not None:
-            self.existing_style.nsfw = self.args.nsfw
-            style_modified = True
-        if self.args.prompt is not None:
-            self.existing_style.prompt = self.args.prompt
-            style_modified = True
-        if self.args.params is not None:
-            self.existing_style.params = self.args.params
-            style_modified = True
-        if len(self.models) > 0:
-            style_modified = True
-        if len(self.tags) > 0:
-            style_modified = True
-        if self.sharedkey is not None:
-            self.existing_style.sharedkey_id = self.sharedkey.id
-            style_modified = True
-        if not style_modified:
-            return {
-                "id": self.existing_style.id,
-                "message": "OK",
-            }, 200
-        db.session.commit()
-        self.existing_style.set_models(self.models)
-        self.existing_style.set_tags(self.tags)
-        return {
-            "id": self.existing_style.id,
-            "message": "OK",
-            "warnings": self.warnings,
-        }, 200
-
-    def validate(self):
-        self.sharedkey = None
-        if self.args.sharedkey:
-            self.shared_key = database.find_sharedkey(self.args.sharedkey)
-            if self.sharedkey is None:
-                raise e.BadRequest("This shared key does not exist", "SharedKeyInvalid")
-            shared_key_validity = self.sharedkey.is_valid()
-            if shared_key_validity[0] is False:
-                raise e.BadRequest(shared_key_validity[1], shared_key_validity[2])
-
-    def delete(self, style_id):
-        self.args = parsers.apikey_parser.parse_args()
-        self.user = database.find_user_by_api_key(self.args["apikey"])
-        if not self.user:
-            raise e.InvalidAPIKey("Style DELETE")
-        if self.user.is_anon():
-            raise e.Forbidden("Anonymous users cannot delete styles", rc="StylesAnonForbidden")
-        self.existing_style = database.get_style_by_uuid(style_id, is_collection=False)
-        if not self.existing_style:
-            raise e.ThingNotFound("Style", style_id)
-        if self.existing_style.user_id != self.user.id and not self.user.moderator:
-            raise e.Forbidden(f"This Style is not owned by user {self.user.get_unique_alias()}")
-        if self.existing_style.user_id != self.user.id and self.user.moderator:
-            logger.info(f"Moderator {self.user.moderator} deleted style {self.existing_style.id}")
-        self.existing_style.delete()
-        return ({"message": "OK"}, 200)
-
-
-## Collections
-
-
-class Collection(Resource):
-    args = None
-
+class WorkerMessages(Resource):
     get_parser = reqparse.RequestParser()
+    get_parser.add_argument("apikey", type=str, required=True, help="User API key.", location="headers")
     get_parser.add_argument(
         "Client-Agent",
         default="unknown:0:unknown",
@@ -3339,11 +3172,23 @@ class Collection(Resource):
         location="headers",
     )
     get_parser.add_argument(
-        "sort",
+        "user_id",
         required=False,
-        default="popular",
         type=str,
-        help="How to sort returned styles. 'popular' sorts by usage and 'age' sorts by date added.",
+        help="Filter the messages by user id (numerical). If user is not moderator, they can only see their own messages.",
+        location="args",
+    )
+    get_parser.add_argument(
+        "worker_id",
+        required=False,
+        type=str,
+        help="Filter the messages by worker id (numerical).",
+        location="args",
+    )
+    get_parser.add_argument(
+        "validity",
+        required=False,
+        type=str,
         location="args",
     )
     get_parser.add_argument(
@@ -3351,325 +3196,181 @@ class Collection(Resource):
         required=False,
         default=1,
         type=int,
-        help="Which page of results to return. Each page has 25 styles.",
-        location="args",
-    )
-    get_parser.add_argument(
-        "type",
-        required=False,
-        default="all",
-        type=str,
-        help="Filter by type. Accepts either 'image', 'text' or 'all'.",
+        help="Which page of results to return. Each page has 50 messages.",
         location="args",
     )
 
-    @cache.cached(timeout=30, query_string=True)
+    # @cache.cached(timeout=60)
     @api.expect(get_parser)
     @api.marshal_with(
-        models.response_model_collection,
+        models.response_model_message_full,
         code=200,
-        description="Lists collection information",
+        description="Worker Message Details",
+        skip_none=True,
         as_list=True,
     )
+    @api.response(401, "Invalid API Key", models.response_model_error)
+    @api.response(403, "Forbidden", models.response_model_error)
+    @api.response(404, "API Key Not Found", models.response_model_error)
     def get(self):
-        """Displays all existing collections. Can filter by type"""
+        """List Worker Messages"""
         self.args = self.get_parser.parse_args()
-        if self.args.sort not in ["popular", "age"]:
-            raise e.BadRequest("'model_state' needs to be one of ['popular', 'age']")
-        if self.args.type not in ["all", "image", "text"]:
-            raise e.BadRequest("'type' needs to be one of ['all', 'image', 'text']")
-        collections = database.retrieve_available_collections(
-            sort=self.args.sort,
-            page=self.args.page - 1,
-            collection_type=self.args.type if self.args.type in ["image", "text"] else None,
+        user = database.find_user_by_api_key(self.args["apikey"])
+        if not user:
+            raise e.InvalidAPIKey("WorkerMessages GET")
+        if not user.moderator and self.args.user_id is None:
+            self.args.user_id = user.id
+        if not user.moderator and user.id != self.args.user_id:
+            raise e.Forbidden("You can only view your own messages.")
+        return (
+            database.get_worker_messages(
+                user_id=self.args.user_id,
+                worker_id=self.args.worker_id,
+                validity=self.args.validity,
+                page=self.args.page - 1,
+            ),
+            200,
         )
-        collections_ret = [co.get_details() for co in collections]
-        return collections_ret, 200
 
     post_parser = reqparse.RequestParser()
-    post_parser.add_argument(
-        "apikey",
-        type=str,
-        required=True,
-        help="The API Key corresponding to a registered user.",
-        location="headers",
-    )
+    post_parser.add_argument("apikey", type=str, required=True, help="User API key.", location="headers")
     post_parser.add_argument(
         "Client-Agent",
         default="unknown:0:unknown",
         type=str,
         required=False,
-        help="The client name and version",
+        help="The client name and version.",
         location="headers",
     )
     post_parser.add_argument(
-        "name",
-        type=str,
-        required=True,
-        location="json",
-    )
-    post_parser.add_argument(
-        "info",
-        type=str,
+        "worker_id",
         required=False,
+        type=str,
+        help="The worker ID to send the message to.",
         location="json",
     )
     post_parser.add_argument(
-        "public",
-        type=bool,
-        default=True,
+        "expiry",
+        type=int,
         required=False,
+        help="The amount of hours from now which this key will stay active. Max 30 days.",
         location="json",
     )
     post_parser.add_argument(
-        "styles",
-        type=list,
+        "message",
         required=True,
+        type=str,
+        help="Filter the messages by worker id (numerical).",
+        location="json",
+    )
+    post_parser.add_argument(
+        "origin",
+        required=False,
+        default="AI Horde Moderators",
+        type=str,
+        help="Set the origin of the message. Moderator only. Non-moderators always set their own ID.",
         location="json",
     )
 
-    decorators = [
-        limiter.limit(
-            limit_value=lim.get_request_90min_limit_per_ip,
-            key_func=lim.get_request_path,
-        ),
-        limiter.limit(limit_value=lim.get_request_2sec_limit_per_ip, key_func=lim.get_request_path),
-    ]
-
-    @api.expect(post_parser, models.input_model_collection, validate=True)
+    @cache.cached(timeout=60)
+    @api.expect(post_parser)
     @api.marshal_with(
-        models.response_model_styles_post,
+        models.response_model_message_full,
         code=200,
-        description="Collection Added",
+        description="Create Worker Message",
         skip_none=True,
     )
-    @api.response(400, "Validation Error", models.response_model_validation_errors)
     @api.response(401, "Invalid API Key", models.response_model_error)
+    @api.response(403, "Forbidden", models.response_model_error)
+    @api.response(404, "API Key Not Found", models.response_model_error)
     def post(self):
-        """Creates a new style collection."""
-        self.warnings = set()
-        # For styles, we just store the models in the params
-        self.styles = []
-        styles_type = None
+        """Create New Worker Message"""
         self.args = self.post_parser.parse_args()
-        if self.args.styles:
-            if len(self.args.styles) < 1:
-                raise e.BadRequest("A collection has to include at least 1 style")
-        else:
-            raise e.BadRequest("A collection has to include at least 1 style")
-        self.user = database.find_user_by_api_key(self.args["apikey"])
-        if not self.user:
-            raise e.InvalidAPIKey("Collection POST")
-        if self.user.is_anon():
-            raise e.Forbidden("Anonymous users cannot create collections", rc="StylesAnonForbidden")
-        for st in self.args.styles:
-            existing_style = database.get_style_by_uuid(st, is_collection=False)
-            if not existing_style:
-                existing_style = database.get_style_by_name(st, is_collection=False)
-                if not existing_style:
-                    raise e.BadRequest(f"A style with name '{st}' cannot be found")
-                if styles_type is None:
-                    styles_type = existing_style.style_type
-                elif styles_type != existing_style.style_type:
-                    raise e.BadRequest("Cannot mix image and text styles in the same collection")
-            self.styles.append(existing_style)
-        self.collection_name = ensure_clean(self.args.name, "collection name")
-        new_collection = StyleCollection(
-            user_id=self.user.id,
-            style_type=styles_type,
-            info=ensure_clean(self.args.info, "collection info") if self.args.info is not None else "",
-            name=self.collection_name,
-            public=self.args.public,
+        user = database.find_user_by_api_key(self.args["apikey"])
+        if not user:
+            raise e.InvalidAPIKey("WorkerMessages POST")
+        if user.is_anon():
+            raise e.AnonForbidden
+        self.origin = self.args.origin
+        if self.args.worker_id is not None:
+            worker = database.find_worker_by_id(self.args.worker_id)
+            if not worker:
+                raise e.WorkerNotFound(self.args.worker_id)
+        if not user.moderator:
+            self.origin = str(user.id)
+            if worker and worker.user_id != self.user.id:
+                raise e.Forbidden("You can only send messages to your own workers.")
+        self.expiry = self.args.expiry
+        # Max expiry is 30 days
+        if self.expiry and self.expiry > 30 * 24:
+            self.expiry = 30 * 24
+        new_message = WorkerMessage(
+            user_id=user.id,
+            worker_id=self.args.worker_id,
+            expiry=datetime.utcnow() + timedelta(hours=self.expiry),
+            message=self.args.message,
+            origin=self.origin,
         )
-        new_collection.create(self.styles)
-        return {
-            "id": new_collection.id,
-            "message": "OK",
-            "warnings": self.warnings,
-        }, 200
+        db.session.add(new_message)
+        db.session.commit()
+        return new_message, 200
 
 
-class SingleCollectionGet(Resource):
-
-    def get_through_id(self, style_id):
-        self.existing_collection = database.get_style_by_uuid(style_id, is_collection=True)
-        if not self.existing_collection:
-            raise e.ThingNotFound("Collection", style_id)
-        return self.existing_collection.get_details()
-
-
-class SingleCollection(SingleCollectionGet):
-    args = None
-
-    @cache.cached(timeout=30, query_string=True)
-    @api.expect(parsers.basic_parser)
-    @api.marshal_with(
-        models.response_model_collection,
-        code=200,
-        description="Lists collection information",
-        as_list=False,
-    )
-    def get(self, collection_id):
-        """Displays information about a single style collection."""
-        return super().get_through_id(collection_id)
-
-    patch_parser = reqparse.RequestParser()
-    patch_parser.add_argument(
-        "apikey",
-        type=str,
-        required=True,
-        help="The API Key corresponding to a registered user.",
-        location="headers",
-    )
-    patch_parser.add_argument(
+class SingleWorkerMessage(Resource):
+    get_parser = reqparse.RequestParser()
+    get_parser.add_argument(
         "Client-Agent",
         default="unknown:0:unknown",
         type=str,
         required=False,
-        help="The client name and version",
+        help="The client name and version.",
         location="headers",
     )
-    patch_parser.add_argument(
-        "name",
-        type=str,
-        required=False,
-        location="json",
-    )
-    patch_parser.add_argument(
-        "info",
-        type=str,
-        required=False,
-        location="json",
-    )
-    patch_parser.add_argument(
-        "public",
-        type=bool,
-        required=False,
-        location="json",
-    )
-    patch_parser.add_argument(
-        "styles",
-        type=list,
-        required=False,
-        location="json",
-    )
 
-    decorators = [
-        limiter.limit(
-            limit_value=lim.get_request_90min_limit_per_ip,
-            key_func=lim.get_request_path,
-        ),
-        limiter.limit(limit_value=lim.get_request_2sec_limit_per_ip, key_func=lim.get_request_path),
-    ]
-
-    @api.expect(patch_parser, models.input_model_collection, validate=True)
+    @cache.cached(timeout=60)
+    @api.expect(get_parser)
     @api.marshal_with(
-        models.response_model_styles_post,
+        models.response_model_message_full,
         code=200,
-        description="Collection Modified",
+        description="Worker Message Details",
         skip_none=True,
     )
-    @api.response(400, "Validation Error", models.response_model_validation_errors)
-    @api.response(401, "Invalid API Key", models.response_model_error)
-    def patch(self, collection_id):
-        """Modifies an existing style collection."""
-        self.warnings = set()
-        # For styles, we just store the models in the params
-        self.styles = []
-        styles_type = None
-        self.args = self.patch_parser.parse_args()
-        if self.args.styles:
-            if len(self.args.styles) < 1:
-                raise e.BadRequest("A collection has to include at least 1 style")
-            for st in self.args.styles:
-                existing_style = database.get_style_by_uuid(st, is_collection=False)
-                if not existing_style:
-                    existing_style = database.get_style_by_name(st, is_collection=False)
-                    if not existing_style:
-                        raise e.BadRequest(f"A style with name '{st}' cannot be found")
-                    if styles_type is None:
-                        styles_type = existing_style.style_type
-                    elif styles_type != existing_style.style_type:
-                        raise e.BadRequest("Cannot mix image and text styles in the same collection", "StyleMismatch")
-                self.styles.append(existing_style)
-        self.user = database.find_user_by_api_key(self.args["apikey"])
-        if not self.user:
-            raise e.InvalidAPIKey("Collection PATCH")
-        self.existing_collection = database.get_style_by_uuid(collection_id, is_collection=True)
-        if not self.existing_collection:
-            raise e.ThingNotFound("Collection", collection_id)
-        if self.existing_collection.user_id != self.user.id:
-            raise e.Forbidden(f"This Collection is not owned by user {self.user.get_unique_alias()}")
-        if self.existing_collection.style_type != styles_type:
-            raise e.BadRequest("Cannot mix image and text styles in the same collection", "StyleMismatch")
-        collection_modified = False
-        if self.args.name:
-            self.existing_collection.name = ensure_clean(self.args.name, "collection name")
-            collection_modified = True
-        if self.args.info is not None:
-            self.existing_collection.info = ensure_clean(self.args.info, "style info")
-            collection_modified = True
-        if self.args.public is not None:
-            self.existing_collection.public = self.args.public
-            collection_modified = True
-        if len(self.styles) > 0:
-            self.existing_collection.styles.clear()
-            for st in self.styles:
-                self.existing_collection.styles.append(st)
-            collection_modified = True
-        if not collection_modified:
-            return {
-                "id": self.existing_collection.id,
-                "message": "OK",
-            }, 200
-        db.session.commit()
-        return {
-            "id": self.existing_collection.id,
-            "message": "OK",
-            "warnings": self.warnings,
-        }, 200
+    def get(self, message_id=""):
+        """Display a Worker Message"""
+        self.args = self.get_parser.parse_args()
+        wmessage = db.session.query(WorkerMessage).filter_by(id=message_id).first()
+        if not wmessage:
+            raise e.ThingNotFound("WorkerMessage", message_id)
+        return wmessage, 200
 
-    @api.expect(parsers.apikey_parser)
+    delete_parser = reqparse.RequestParser()
+    delete_parser.add_argument("apikey", type=str, required=True, help="User API key.", location="headers")
+    delete_parser.add_argument(
+        "Client-Agent",
+        default="unknown:0:unknown",
+        type=str,
+        required=False,
+        help="The client name and version.",
+        location="headers",
+    )
+
+    @api.expect(delete_parser)
     @api.marshal_with(
         models.response_model_simple_response,
         code=200,
-        description="Operation Completed",
-        skip_none=True,
+        description="Delete Worker Message",
     )
-    @api.response(400, "Validation Error", models.response_model_validation_errors)
-    @api.response(401, "Invalid API Key", models.response_model_error)
-    def delete(self, collection_id):
-        """Deletes a style collection."""
-        self.args = parsers.apikey_parser.parse_args()
-        self.user = database.find_user_by_api_key(self.args["apikey"])
-        if not self.user:
-            raise e.InvalidAPIKey("Collection PATCH")
-        self.existing_collection = database.get_style_by_uuid(collection_id, is_collection=True)
-        if not self.existing_collection:
-            raise e.ThingNotFound("Collection", collection_id)
-        if self.existing_collection.user_id != self.user.id and not self.user.moderator:
-            raise e.Forbidden(f"This Collection is not owned by user {self.user.get_unique_alias()}")
-        if self.existing_collection.user_id != self.user.id and self.user.moderator:
-            logger.info(f"Moderator {self.user.moderator} deleted collection {self.existing_collection.id}")
-        self.existing_collection.delete()
-        return ({"message": "OK"}, 200)
-
-
-class SingleCollectionByName(SingleCollectionGet):
-    @cache.cached(timeout=30)
-    @api.expect(parsers.basic_parser)
-    @api.marshal_with(
-        models.response_model_collection,
-        code=200,
-        description="Lists collection information by name",
-        as_list=False,
-    )
-    def get(self, collection_name):
-        """Seeks an style collection by name and displays its information."""
-        self.existing_collection = database.get_style_by_name(collection_name)
-        if not self.existing_collection:
-            raise e.ThingNotFound("Collection", collection_name)
-        return self.existing_collection.get_details()
-
-
-# TODO: vote and transfer kudos on vote
+    def delete(self, message_id=""):
+        """Delete a Worker Message"""
+        self.args = self.delete_parser.parse_args()
+        user = database.find_user_by_api_key(self.args["apikey"])
+        if not user:
+            raise e.InvalidAPIKey("WorkerMessages DELETE")
+        wmessage = db.session.query(WorkerMessage).filter_by(id=message_id).first()
+        if not wmessage:
+            raise e.ThingNotFound("WorkerMessage", message_id)
+        if wmessage.user_id != user.id and user.moderator is False:
+            raise e.Forbidden("You can only delete your own messages.")
+        db.session.delete(wmessage)
+        db.session.commit()
+        return {"message": "OK"}, 200
